Moved some things around a bit, stopped trying to manage the other projects we depend on (packer-builder-arm-image and pkgbuilder).
Now fully support docker, so you can spin up a dockerized version of the
cluster, and run a Nomad client on the same VM running the cluster
containers.  Ideally it would've been preferable to run the nomad jobs
within a container instead, but I didn't want to deal with docker-in-docker,
and even just sharing the host's docker daemon didn't pan out.