We don't really care about consul data on client nodes, so blow it away on startup.  This is in response to a case where corruption prevented consul coming up without intervention.
Retry startup basically indefinitely, and don't let quick exits prevent
restarts.  This was motivated by a case where after a power failure the
network didn't come up as quickly as consul and nomad did, so we didn't
have an interface to bind to.  In general it seems desirable to keep
trying even on failure, to reduce the need for intervention.  We should
make a better log rotation strategy to avoid this causing its own problems.